Move seed file content to script/seed.js, so we can run it with npm run seed
